refactor(object_store): Polish implementation details of object_store
